requirements.txt
#Dependencies
numpy 
streamlit
python-dotenv
google-generativeai
pillow 